Move beak project code around
