+++ packages/electron-host/src/main.ts
@@ -7,9 +7,9 @@ import { autoUpdater } from 'electron-updater';
 import './ipc-layer';
 import './updater';
 import arbiter from './lib/arbiter';
-import { tryOpenProjectFolder } from './lib/beak-project';
 import nestClient from './lib/nest-client';
 import persistentStore from './lib/persistent-store';
+import { tryOpenProjectFolder } from './lib/project';
 import handleUrlEvent from './protocol';
 import { createAndSetMenu } from './utils/menu';
 import { appIsPackaged } from './utils/static-path';
